chore: rename history types file
