+++ src/History/history_helpers.ts
@@ -1,5 +1,5 @@
 import { OrteliusUTXO } from '@/Explorer';
-import { iHistoryNftFamilyBalance } from '@/History/parsed_types';
+import { iHistoryNftFamilyBalance } from '@/History/types';
 import { AVMConstants } from 'avalanche/dist/apis/avm';
 import { parseNftPayload } from '@/utils';
 import { isOutputOwner } from '@/Explorer/ortelius/utxoUtils';
