+++ src/History/index.ts
@@ -1,4 +1,4 @@
 export * from './history_helpers';
 export * from './parsers';
 export * from './evmParser';
-export * from './parsed_types';
+export * from './types';
